6.12 implemented
Notification working
