+++ Exercises/Part-6/redux-anecdotes/src/App.jsx
@@ -1,11 +1,13 @@
 import AnecdoteForm from './components/AnecdoteForm'
 import { AnecdoteList } from './components/AnecdoteList'
 import { Filter } from './components/Filter'
+import Notification from './components/Notification'
 
 const App = () => {
 
   return (
     <div>
+      <Notification />
       <h2>Anecdotes</h2>
       <Filter />
       <AnecdoteList />
