--- Exercises/Part-6/redux-anecdotes/src/components/Notification.js
@@ -1,14 +1,0 @@
-const Notification = () => {
-  const style = {
-    border: 'solid',
-    padding: 10,
-    borderWidth: 1
-  }
-  return (
-    <div style={style}>
-      render here notification...
-    </div>
-  )
-}
-
-export default Notification
