+++ Exercises/Part-6/redux-anecdotes/src/reducers/anecdoteReducer.js
@@ -44,7 +44,7 @@ const anecdoteSlice = createSlice({
       return state.map(anecdote => 
         anecdote.id !== id ? anecdote : changeAnecdote
       )
-    }
+    },
   }
 })
