+++ Exercises/Part-6/redux-anecdotes/src/reducers/notificationReducer.jsx
@@ -1,0 +1,16 @@
+import { createSlice } from "@reduxjs/toolkit";
+
+const initialState = 'works'
+
+const notificationSlice = createSlice({
+    name: 'notification',
+    initialState,
+    reducers: {
+        notificationAction(state, action) {
+            return action.payload
+        }
+    }
+})
+
+export default notificationSlice.reducer
+export const { notificationAction} = notificationSlice.actions
