+++ Exercises/Part-6/redux-anecdotes/src/store.js
@@ -1,11 +1,13 @@
 import { configureStore } from '@reduxjs/toolkit'
 import filterReducer from './reducers/filterReducer'
 import anecdoteReducer from './reducers/anecdoteReducer'
+import notificationReducer from './reducers/notificationReducer'
 
 const store = configureStore({
     reducer: {
         anecdotes: anecdoteReducer,
-        filter: filterReducer
+        filter: filterReducer,
+        notification: notificationReducer
     }
 })
